refactor: move anr detection code
